backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.2.0

# AI and ML
google-generativeai==0.3.2
sentence-transformers==2.2.2
torch==2.1.0

# Vector Database
pinecone-client==2.2.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Database (if using Convex)
convex==0.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0